move NativeError.kotlinCause from main to test
