+++ kmp-nativecoroutines-core/src/appleTest/kotlin/com/rickclephas/kmp/nativecoroutines/NSErrorTests.kt
@@ -2,20 +2,12 @@ package com.rickclephas.kmp.nativecoroutines
 
 import kotlinx.cinterop.UnsafeNumber
 import kotlinx.cinterop.convert
-import kotlin.native.concurrent.isFrozen
 import kotlin.test.*
 
+internal actual val NativeError.kotlinCause
+    get() = this.userInfo["KotlinException"] as? Throwable
+
 class NSErrorTests {
-
-    @Test
-    fun `ensure frozen`() {
-        val exception = RandomException()
-        assertFalse(exception.isFrozen, "Exception shouldn't be frozen yet")
-        val nsError = exception.asNativeError()
-        assertTrue(nsError.isFrozen, "NSError should be frozen")
-        assertTrue(exception.isFrozen, "Exception should be frozen")
-    }
-
     @Test
     @OptIn(UnsafeNumber::class)
     fun `ensure NSError domain and code are correct`() {
@@ -31,12 +23,5 @@ class NSErrorTests {
         val nsError = exception.asNativeError()
         assertEquals(exception.message, nsError.localizedDescription,
             "Localized description isn't set to message")
-    }
-
-    @Test
-    fun `ensure exception is part of user info`() {
-        val exception = RandomException()
-        val nsError = exception.asNativeError()
-        assertSame(exception, nsError.kotlinCause, "Exception isn't part of the user info")
     }
 }
